This fixes 'val' not working in javac7. The LogDisabler used to simply suppress writes to standard out, but javac7 has a redesigned error system where errors are cached. Now the LogDisabler also wipes out any cached errors added in between a 'disable' and 'enable' session. Ironically we actually whined to Joe Darcy about javac6 emitting errors 'too early' and this caching system is the result. Heh.
This was already the case, but a disableLoggers() call HAS to be paired with an enableLoggers() call, so use try/finally.

Fixes issue #259.

+++ src/core/lombok/javac/handlers/HandleVal.java
@@ -72,7 +72,7 @@ public class HandleVal extends JavacASTAdapter {
 			return;
 		}
 		
-		JavacHandlerUtil.deleteImportFromCompilationUnit(localNode, "lombok.val");
+		if (localNode.shouldDeleteLombokAnnotations()) JavacHandlerUtil.deleteImportFromCompilationUnit(localNode, "lombok.val");
 		
 		local.mods.flags |= Flags.FINAL;
 		
@@ -115,7 +115,7 @@ public class HandleVal extends JavacASTAdapter {
 				if (replacement != null) {
 					local.vartype = replacement;
 				} else {
-					local.vartype = JavacResolution.createJavaLangObject(localNode.getTreeMaker(), localNode.getAst());;
+					local.vartype = JavacResolution.createJavaLangObject(localNode.getTreeMaker(), localNode.getAst());
 				}
 				localNode.getAst().setChanged();
 			} catch (JavacResolution.TypeNotConvertibleException e) {
